chore: 평점 자료형 변경

+++ src/main/java/com/uhdyl/backend/review/dto/request/ReviewCreateRequest.java
@@ -2,7 +2,7 @@ package com.uhdyl.backend.review.dto.request;
 
 public record ReviewCreateRequest (
         String content,
-        Long rating,
+        Double rating,
         String imageUrl,
         String publicId,
         Long productId
