+++ src/main/java/com/uhdyl/backend/review/dto/response/ReviewResponse.java
@@ -4,11 +4,10 @@ import com.uhdyl.backend.review.domain.Review;
 
 import java.time.LocalDateTime;
 
-// TODO: 상품 도메인 개발 후 상품의 제목도 함께 전달하기
 public record ReviewResponse(
         Long Id,
         String content,
-        Long rating,
+        Double rating,
         String nickName,
         String imageUrl,
         String title,
